refactor: update `require_path` in `gemspec` for each of the gems (#106)
