+++ src/packages/axe-core-capybara/spec/axe-capybara_spec.rb
@@ -1,18 +1,18 @@
 require "spec_helper"
-require "tempfile"
+require "capybara/rspec"
 
-require_relative "../axe-core-selenium"
+require_relative "../lib/axe-capybara"
 
-describe AxeSelenium do
+describe AxeCapybara do
   subject { described_class }
 
   describe "driver" do
     it "validate yielded configuration" do
-      driver = AxeSelenium.configure("firefox") do
+      driver = AxeCapybara.configure("firefox") do
       end
 
       expect(driver).not_to be_nil
-      driver.page.navigate.to "https://google.com" # can navigate
+      driver.page.visit "https://google.com" # can navigate
       expect(driver).to respond_to :skip_iframes # can config
       expect(driver).to respond_to :jslib
       expect(driver.jslib).to include("axe.run=function(") # has axe injected
@@ -34,7 +34,7 @@ describe AxeSelenium do
       # configure:
       # 1. driver for browser
       # 2. a different js path
-      AxeSelenium.configure do |c|
+      AxeCapybara.configure do |c|
         c.jslib_path = different_axe_path
       end
 
@@ -43,25 +43,16 @@ describe AxeSelenium do
     end
 
     it "should yield configuration with Safari driver" do
-      AxeSelenium.configure("safari") do
+      AxeCapybara.configure("safari") do
       end
 
       actual = Axe::Configuration.instance
       expect(actual.page).not_to be_nil
-      expect(actual.page.to_s).to include("WebDriver::Safari")
-    end
-
-    it "should yield configuration with Firefox driver" do
-      AxeSelenium.configure("firefox") do
-      end
-
-      actual = Axe::Configuration.instance
-      expect(actual.page).not_to be_nil
-      expect(actual.page.to_s).to include("WebDriver::Firefox")
+      expect(actual.page.to_s).to include("Capybara::Selenium::Driver")
     end
 
     it "should raise when no configuration block is provided" do
-      expect { AxeSelenium.configure }.to raise_error("Please provide a configure block for AxeSelenium")
+      expect { AxeCapybara.configure }.to raise_error("Please provide a configure block for AxeCapybara")
     end
   end
 end
