+++ src/packages/axe-core-rspec/axe-core-rspec.gemspec
@@ -6,9 +6,12 @@ Gem::Specification.new do |spec|
   spec.authors = ["Deque Systems"]
   spec.platform = Gem::Platform::RUBY
 
+  spec.require_path = "lib"
+
   spec.add_dependency "dumb_delegator"
   spec.add_dependency "virtus"
 
-  spec.add_development_dependency "rake", "~> 12.0"
-  spec.add_development_dependency "rspec", "~> 3.6"
+  spec.add_development_dependency "rake"
+  spec.add_development_dependency "rspec"
+  spec.add_development_dependency "rspec-its"
 end
