+++ src/packages/axe-core-rspec/lib/axe-rspec.rb
@@ -1,6 +1,6 @@
 require "rspec/core"
 
-require_relative "../../common/axe/matchers/be_accessible"
+require_relative "../../../common/axe/matchers/be_accessible"
 
 RSpec.configure do |config|
   config.include Axe::Matchers
