+++ src/packages/axe-core-selenium/spec/axe-selenium_spec.rb
@@ -1,17 +1,18 @@
 require "spec_helper"
-require "capybara/rspec"
-require_relative "../axe-core-capybara"
+require "tempfile"
 
-describe AxeCapybara do
+require_relative "../lib/axe-selenium"
+
+describe AxeSelenium do
   subject { described_class }
 
   describe "driver" do
     it "validate yielded configuration" do
-      driver = AxeCapybara.configure("firefox") do
+      driver = AxeSelenium.configure("firefox") do
       end
 
       expect(driver).not_to be_nil
-      driver.page.visit "https://google.com" # can navigate
+      driver.page.navigate.to "https://google.com" # can navigate
       expect(driver).to respond_to :skip_iframes # can config
       expect(driver).to respond_to :jslib
       expect(driver.jslib).to include("axe.run=function(") # has axe injected
@@ -33,7 +34,7 @@ describe AxeCapybara do
       # configure:
       # 1. driver for browser
       # 2. a different js path
-      AxeCapybara.configure do |c|
+      AxeSelenium.configure do |c|
         c.jslib_path = different_axe_path
       end
 
@@ -42,16 +43,25 @@ describe AxeCapybara do
     end
 
     it "should yield configuration with Safari driver" do
-      AxeCapybara.configure("safari") do
+      AxeSelenium.configure("safari") do
       end
 
       actual = Axe::Configuration.instance
       expect(actual.page).not_to be_nil
-      expect(actual.page.to_s).to include("Capybara::Selenium::Driver")
+      expect(actual.page.to_s).to include("WebDriver::Safari")
+    end
+
+    it "should yield configuration with Firefox driver" do
+      AxeSelenium.configure("firefox") do
+      end
+
+      actual = Axe::Configuration.instance
+      expect(actual.page).not_to be_nil
+      expect(actual.page.to_s).to include("WebDriver::Firefox")
     end
 
     it "should raise when no configuration block is provided" do
-      expect { AxeCapybara.configure }.to raise_error("Please provide a configure block for AxeCapybara")
+      expect { AxeSelenium.configure }.to raise_error("Please provide a configure block for AxeSelenium")
     end
   end
 end
