+++ src/packages/axe-core-watir/lib/axe-watir.rb
@@ -1,7 +1,7 @@
-require "selenium-webdriver"
-require_relative "../../common/axe/configuration"
+require "watir"
+require_relative "../../../common/axe/configuration"
 
-module AxeSelenium
+module AxeWatir
 
   # configure method
   # - which takes an optional argument browser
@@ -10,12 +10,12 @@ module AxeSelenium
 
     # instantiate axe configuration (singleton) with defaults or given config
     if !block_given?
-      raise Exception.new "Please provide a configure block for AxeSelenium"
+      raise Exception.new "Please provide a configure block for AxeWatir"
     end
 
     config = Axe::Configuration.instance
 
-    # provide a selenium webdriver page object
+    # provide a watir webdriver page object
     config.page = get_driver(browser)
 
     # await and return
@@ -29,13 +29,17 @@ module AxeSelenium
   def self.get_driver(browserName)
     case browserName
     when "chrome"
-      Selenium::WebDriver.for :chrome
+      Watir::Browser.new :chrome
     when "safari"
-      Selenium::WebDriver.for :safari
+      Watir::Browser.new :safari
+    when "internet_explorer"
+      Watir::Browser.new :internet_explorer
+    when "edge"
+      Watir::Browser.new :edge
     when "firefox"
-      Selenium::WebDriver.for :firefox
+      Watir::Browser.new :firefox
     else
-      Selenium::WebDriver.for :firefox
+      Watir::Browser.new :firefox
     end
   end
 end
